# REST-API-mono/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Real Estate API",
    "description": "API для управління квартирами та користувачами"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Реєстрація нового користувача",
        "description": "Реєстрація нового користувача з електронною адресою, паролем та роллю.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "Електронна адреса користувача"
                },
                "password": {
                  "type": "string",
                  "description": "Пароль користувача"
                },
                "role": {
                  "type": "string",
                  "description": "Роль користувача (owner, realtor, buyer)"
                }
              },
              "required": [
                "email",
                "password",
                "role"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Користувач успішно зареєстрований",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Повідомлення про успішну реєстрацію"
                },
                "token": {
                  "type": "string",
                  "description": "JWT токен для автентифікації"
                }
              }
            }
          },
          "400": {
            "description": "Помилка при реєстрації користувача"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Автентифікація користувача",
        "description": "Автентифікація користувача за допомогою електронної адреси та пароля.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "description": "Електронна адреса користувача"
                },
                "password": {
                  "type": "string",
                  "description": "Пароль користувача"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Автентифікація успішна",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Повідомлення про успішну автентифікацію"
                },
                "token": {
                  "type": "string",
                  "description": "JWT токен для автентифікації"
                }
              }
            }
          },
          "401": {
            "description": "Неправильна електронна адреса або пароль"
          }
        }
      }
    },
    "/apartments": {
      "get": {
        "summary": "Отримати список квартир",
        "description": "Отримати список всіх квартир з можливістю фільтрації.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "type": "string",
            "description": "Розташування квартири"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "type": "integer",
            "description": "Максимальна ціна квартири"
          },
          {
            "name": "minRooms",
            "in": "query",
            "type": "integer",
            "description": "Мінімальна кількість кімнат у квартирі"
          },
          {
            "name": "isFeatured",
            "in": "query",
            "type": "boolean",
            "description": "Фільтрувати за прапорцем \"isFeatured\""
          },
          {
            "name": "isTop",
            "in": "query",
            "type": "boolean",
            "description": "Фільтрувати за прапорцем \"isTop\""
          }
        ],
        "responses": {
          "200": {
            "description": "Список квартир",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Apartment"
              }
            }
          },
          "400": {
            "description": "Помилка при отриманні списку квартир"
          }
        }
      },
      "post": {
        "summary": "Створити нову квартиру",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Створити нову квартиру з вказаними даними.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewApartment"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Квартира успішно створена",
            "schema": {
              "$ref": "#/definitions/Apartment"
            }
          },
          "400": {
            "description": "Помилка при створенні квартири"
          }
        }
      }
    },
    "/apartments/{id}": {
      "get": {
        "summary": "Отримати інформацію про квартиру",
        "description": "Отримати інформацію про квартиру за її ідентифікатором.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Ідентифікатор квартири"
          }
        ],
        "responses": {
          "200": {
            "description": "Інформація про квартиру",
            "schema": {
              "$ref": "#/definitions/Apartment"
            }
          },
          "404": {
            "description": "Квартира не знайдена"
          },
          "400": {
            "description": "Помилка при отриманні інформації про квартиру"
          }
        }
      },
      "put": {
        "summary": "Оновити інформацію про квартиру",
        "description": "Оновити інформацію про квартиру за її ідентифікатором.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Ідентифікатор квартири"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateApartment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Інформація про оновлену квартиру",
            "schema": {
              "$ref": "#/definitions/Apartment"
            }
          },
          "404": {
            "description": "Квартира не знайдена"
          },
          "400": {
            "description": "Помилка при оновленні інформації про квартиру"
          }
        }
      },
      "delete": {
        "summary": "Видалити квартиру",
        "description": "Видалити квартиру за її ідентифікатором.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Ідентифікатор квартири"
          }
        ],
        "responses": {
          "200": {
            "description": "Квартира успішно видалена"
          },
          "404": {
            "description": "Квартира не знайдена"
          },
          "400": {
            "description": "Помилка при видаленні квартири"
          }
        }
      }
    }
  },
  "definitions": {
    "Apartment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "Ідентифікатор квартири"
        },
        "address": {
          "type": "string",
          "description": "Адреса квартири"
        },
        "price": {
          "type": "integer",
          "description": "Ціна квартири"
        },
        "rooms": {
          "type": "integer",
          "description": "Кількість кімнат у квартирі"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Масив посилань на зображення квартири"
        },
        "ownerEmail": {
          "type": "string",
          "description": "Електронна адреса власника квартири"
        },
        "views": {
          "type": "integer",
          "description": "Кількість переглядів квартири"
        },
        "isFeatured": {
          "type": "boolean",
          "description": "Прапорець, що вказує на те, чи є квартира у списку рекомендованих"
        },
        "isTop": {
          "type": "boolean",
          "description": "Прапорець, що вказує на те, чи є квартира в топі оголошень"
        }
      }
    },
    "NewApartment": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "description": "Адреса квартири"
        },
        "price": {
          "type": "integer",
          "description": "Ціна квартири"
        },
        "rooms": {
          "type": "integer",
          "description": "Кількість кімнат у квартирі"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Масив посилань на зображення квартири"
        },
        "ownerEmail": {
          "type": "string",
          "description": "Електронна адреса власника квартири"
        },
        "isFeatured": {
          "type": "boolean",
          "description": "Прапорець, що вказує на те, чи є квартира у списку рекомендованих"
        },
        "isTop": {
          "type": "boolean",
          "description": "Прапорець, що вказує на те, чи є квартира в топі оголошень"
        }
      },
      "required": [
        "address",
        "price",
        "rooms",
        "ownerEmail"
      ]
    },
    "UpdateApartment": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "description": "Нова адреса квартири"
        },
        "price": {
          "type": "integer",
          "description": "Нова ціна квартири"
        },
        "rooms": {
          "type": "integer",
          "description": "Нова кількість кімнат у квартирі"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "Новий масив посилань на зображення квартири"
        },
        "ownerEmail": {
          "type": "string",
          "description": "Нова електронна адреса власника квартири"
        },
        "isFeatured": {
          "type": "boolean",
          "description": "Нове значення для прапорця, що вказує на те, чи є квартира у списку рекомендованих"
        },
        "isTop": {
          "type": "boolean",
          "description": "Нове значення для прапорця, що вказує на те, чи є квартира в топі оголошень"
        }
      }
    }
  }
}